Set up cross-platform ModDLL loading
