+++ Sources/Plasma/Apps/plClient/main.cpp
@@ -50,8 +50,6 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 void plClient::IResizeNativeDisplayDevice(int width, int height, bool windowed) {}
 void plClient::IChangeResolution(int width, int height) {}
 void plClient::IUpdateProgressIndicator(plOperationProgress* progress) {}
-void plClient::InitDLLs() {}
-void plClient::ShutdownDLLs() {}
 void plClient::ShowClientWindow() {}
 void plClient::FlashWindow() {}
