+++ Sources/Plasma/CoreLib/HeadSpin.h
@@ -71,14 +71,17 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
     typedef HWND hsWindowHndl;
     typedef HINSTANCE hsWindowInst;
     typedef HINSTANCE HMODULE;
+    typedef HMODULE hsLibraryHndl;
     typedef long HRESULT;
     typedef void* HANDLE;
 #elif HS_BUILD_FOR_MACOS
     typedef void* hsWindowHndl;
     typedef void* hsWindowInst;
+    typedef void* hsLibraryHndl;
 #else
     typedef int32_t* hsWindowHndl;
     typedef int32_t* hsWindowInst;
+    typedef void* hsLibraryHndl;
 #endif // HS_BUILD_FOR_WIN32
 
 //======================================
